Rename App's keyword
