+++ app/src/main/res/drawable/avd_heart_fill.xml
@@ -16,7 +16,7 @@ Copyright (C) 2015 The Android Open Source Project
 -->
 <animated-vector
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:drawable="@drawable/ic_heart">
+    android:drawable="@drawable/ic_heart_2">
 
     <target
         android:name="clip"
