+++ app/src/main/res/drawable/ic_heart.xml
@@ -1,38 +1,9 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-Copyright (C) 2015 The Android Open Source Project
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-         http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<vector
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="280dp"
-    android:height="280dp"
-    android:viewportWidth="56"
-    android:viewportHeight="56">
-
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="52dp"
+    android:height="52dp"
+    android:viewportWidth="51.997"
+    android:viewportHeight="51.997">
     <path
-        android:name="empty"
-        android:pathData="@string/heart_empty_path"
-        android:fillColor="@android:color/holo_red_light" />
-
-    <clip-path
-        android:name="clip"
-        android:pathData="@string/heart_clip_hidden" />
-
-    <path
-        android:name="full"
-        android:pathData="@string/heart_full_path"
-        android:fillColor="@android:color/holo_red_light"/>
-
+        android:fillColor="#FF000000"
+        android:pathData="M51.911,16.242C51.152,7.888 45.239,1.827 37.839,1.827c-4.93,0 -9.444,2.653 -11.984,6.905c-2.517,-4.307 -6.846,-6.906 -11.697,-6.906c-7.399,0 -13.313,6.061 -14.071,14.415c-0.06,0.369 -0.306,2.311 0.442,5.478c1.078,4.568 3.568,8.723 7.199,12.013l18.115,16.439l18.426,-16.438c3.631,-3.291 6.121,-7.445 7.199,-12.014C52.216,18.553 51.97,16.611 51.911,16.242zM49.521,21.261c-0.984,4.172 -3.265,7.973 -6.59,10.985L25.855,47.481L9.072,32.25c-3.331,-3.018 -5.611,-6.818 -6.596,-10.99c-0.708,-2.997 -0.417,-4.69 -0.416,-4.701l0.015,-0.101C2.725,9.139 7.806,3.826 14.158,3.826c4.687,0 8.813,2.88 10.771,7.515l0.921,2.183l0.921,-2.183c1.927,-4.564 6.271,-7.514 11.069,-7.514c6.351,0 11.433,5.313 12.096,12.727C49.938,16.57 50.229,18.264 49.521,21.261z" />
 </vector>
