+++ app/src/main/res/drawable/ic_search.xml
@@ -4,6 +4,6 @@
     android:viewportWidth="512.392"
     android:viewportHeight="512.392">
   <path
-      android:fillColor="#FF000000"
+      android:fillColor="@android:color/darker_gray"
       android:pathData="m211.196,422c-116.346,0 -211,-94.654 -211,-211s94.654,-211 211,-211 211,94.654 211,211 -94.654,211 -211,211zM211.196,40c-94.29,0 -171,76.71 -171,171s76.71,171 171,171 171,-76.71 171,-171 -76.71,-171 -171,-171zM506.339,506.534c7.81,-7.811 7.81,-20.475 0,-28.285l-89.5,-89.5c-7.811,-7.811 -20.475,-7.811 -28.285,0s-7.81,20.475 0,28.285l89.5,89.5c3.905,3.905 9.024,5.857 14.143,5.857s10.236,-1.952 14.142,-5.857z"/>
 </vector>
