+++ app/src/main/res/layout/fragment_home.xml
@@ -1,33 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools">
+        android:layout_height="match_parent">
 
         <TextView
             android:id="@+id/avd_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle1"
             android:layout_margin="24dp"
-            app:layout_constraintTop_toTopOf="parent"
+            style="@style/TextAppearance.Title"
             app:layout_constraintStart_toStartOf="parent"
-            tools:text="Player Drawable"/>
+            app:layout_constraintTop_toTopOf="parent"
+            tools:text="Player Drawable" />
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/avd_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:overScrollMode="never"
             android:layout_margin="12dp"
+            android:overScrollMode="never"
             app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-            app:spanCount="3"
-            app:layout_constraintTop_toBottomOf="@id/avd_title"
             app:layout_constraintBottom_toBottomOf="parent"
-            tools:listitem="@layout/item_avd"
-            tools:itemCount="8"/>
+            app:layout_constraintTop_toBottomOf="@id/avd_title"
+            app:spanCount="2"
+            tools:itemCount="8"
+            tools:listitem="@layout/item_avd" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
