+++ app/src/main/res/navigation/main_navigation.xml
@@ -3,13 +3,13 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_common">
+    app:startDestination="@+id/navigation_home">
 
     <fragment
-        android:id="@+id/navigation_common"
-        android:name="com.ggu.avd.ui.list.ListFragment"
-        android:label="@string/title_common"
-        tools:layout="@layout/fragment_list">
+        android:id="@+id/navigation_home"
+        android:name="com.ggu.avd.ui.home.HomeFragment"
+        android:label="@string/title_home"
+        tools:layout="@layout/fragment_home">
 
         <argument
             android:name="avdType"
@@ -18,10 +18,10 @@
     </fragment>
 
     <fragment
-        android:id="@+id/navigation_player"
-        android:name="com.ggu.avd.ui.list.ListFragment"
-        android:label="@string/title_player"
-        tools:layout="@layout/fragment_list">
+        android:id="@+id/navigation_search"
+        android:name="com.ggu.avd.ui.search.SearchFragment"
+        android:label="@string/title_search"
+        tools:layout="@layout/fragment_search">
         <argument
             android:name="avdType"
             android:defaultValue="1"
@@ -34,7 +34,7 @@
         android:label="@string/title_my"
         tools:layout="@layout/fragment_my">
         <action
-            android:id="@+id/action_my_fragment_to_common_fragment"
-            app:destination="@id/navigation_common" />
+            android:id="@+id/action_my_fragment_to_home_fragment"
+            app:destination="@id/navigation_home" />
     </fragment>
 </navigation>
