+++ app/src/main/res/values/strings.xml
@@ -2,8 +2,12 @@
     <string name="app_name">AVD Test</string>
 
     <string name="title_my">My</string>
-    <string name="title_common">Common</string>
+    <string name="title_home">Home</string>
+    <string name="title_search">Search</string>
     <string name="title_player">Player</string>
+
+    <string name="title_search_keyword">Keywords</string>
+    <string name="title_search_result">Result</string>
 
     <string name="my_empty">My Drawable is empty</string>
     <string name="btn_add">Add Drawable</string>
